Fix schema type issue

+++ src/food/queries/FoodDetailContainerQuery.js
@@ -1,7 +1,7 @@
 import { graphql } from 'react-relay';
 
 const FoodDetailContainerQuery = graphql`
-  query FoodDetailContainerQuery($foodId: Float!) {
+  query FoodDetailContainerQuery($foodId: BigFloat!) {
     foodByRowId(rowId: $foodId) {
       ...FoodDetailContainer_foodByRowId,
     }
