+++ src/food/queries/FoodSearchContainerQuery.js
@@ -1,7 +1,7 @@
 import { graphql } from 'react-relay';
 
 const FoodSearchContainerQuery = graphql`
-  query FoodSearchContainerQuery($foodId: Float, $foodDescription: String) {
+  query FoodSearchContainerQuery($foodId: BigFloat, $foodDescription: String) {
     searchFoods(foodId: $foodId, foodDescription: $foodDescription) {
       ...FoodSearchContainer_searchFoods,
     }
